Generate a requirements file from a packrat.lock file

Generate a requirements file from a packrat.lock file

rip_freeze_packrat(packrat_lock_path = "packrat/packrat.lock",
  output_path = "requirements.txt", eq_sym = ">=", append = TRUE)

Arguments

packrat_lock_path
Path to packrat.lock to extract requirements from.

output_path
String indicating where to write resulting requirements file to.

eq_sym
The equality symbol to be used when writing requirements (i.e. package>=1.0.0). Use NULL to not include package versions in your requirements file.

append
Should requirements from packrat_lock_path be appended to end of output_path? Overwrites contents of output_path if FALSE.

Value

Nothing is returned. Results are written to output_path

Examples

if (FALSE) {
rip_freeze_packrat(output_path = "my_requirements.txt")
rip_freeze_packrat(output_path = "versionless_requirements.txt", eq_sym = NULL)
rip_freeze_packrat(output_path = "overwrite_requirements.txt", append = FALSE)
}